pyproject.toml
[project]
name = "vaudeville"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mlx-lm>=0.21.0",
    "pyyaml>=6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vaudeville"]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "types-pyyaml>=6.0.12.20250915",
]
